Rename everything tcs34725 to colour
