+++ sense_hat/__init__.py
@@ -12,5 +12,5 @@ from .stick import (
     ACTION_RELEASED,
     ACTION_HELD,
     )
-from .tcs34725 import TCS34725, TCS34725 as ColourSensor
+from .colour import ColourSensor
 __version__ = '2.3.0'
